Test with Quick and Nimble
Why:

* Testing with XCTest can be verbose and testing async behavior is a mess.

This change addresses the need by:

* Using Quick and Nimble for testing
