+++ Cartfile.private
@@ -1,0 +1,2 @@
+github "Quick/Quick"
+github "Quick/Nimble"
