+++ Cartfile.resolved
@@ -1,0 +1,2 @@
+github "Quick/Nimble" "v3.0.0"
+github "Quick/Quick" "v0.8.0"
